chore: rename types dir to models
