+++ Frontend-nextjs/app/(protected)/newRicorrenza/NewRicorrenzaModal.tsx
@@ -9,7 +9,7 @@ import Dialog from "@/app/components/ui/Dialog";
 import LoadingOverlay from "@/app/components/ui/LoadingOverlay";
 import NewRicorrenzaForm from "./NewRicorrenzaForm";
 import { useCategories } from "@/context/contexts/CategoriesContext";
-import { Ricorrenza, RicorrenzaBase } from "@/types/types/ricorrenza";
+import { Ricorrenza, RicorrenzaBase } from "@/types/models/ricorrenza";
 import type { NewRicorrenzaModalProps } from "@/types";
 
 // ============================
