+++ Frontend-nextjs/types/index.ts
@@ -1,8 +1,8 @@
-export * from "./types/sidebar";
-export * from "./types/transaction";
-export * from "./types/category";
-export * from "./types/ricorrenza";
-export * from "./types/user";
+export * from "./models/sidebar";
+export * from "./models/transaction";
+export * from "./models/category";
+export * from "./models/ricorrenza";
+export * from "./models/user";
 export * from "./profilo";
 export * from "./transazioni";
 export * from "./ricorrenti";
