+++ Frontend-nextjs/types/newRicorrenza/index.ts
@@ -1,18 +1,18 @@
 export type NewRicorrenzaModalProps = {
     open: boolean;
     onClose: () => void;
-    ricorrenzaToEdit?: import("@/types/types/ricorrenza").Ricorrenza | null;
-    onSave: (data: import("@/types/types/ricorrenza").RicorrenzaBase) => Promise<void>;
+    ricorrenzaToEdit?: import("@/types/models/ricorrenza").Ricorrenza | null;
+    onSave: (data: import("@/types/models/ricorrenza").RicorrenzaBase) => Promise<void>;
 };
 
 export type NewRicorrenzaFormProps = {
-    onSave: (data: import("@/types/types/ricorrenza").RicorrenzaBase) => Promise<void>;
+    onSave: (data: import("@/types/models/ricorrenza").RicorrenzaBase) => Promise<void>;
     onCancel: () => void;
-    initialValues?: Partial<import("@/types/types/ricorrenza").RicorrenzaBase>;
-    onChangeForm?: (data: Partial<import("@/types/types/ricorrenza").RicorrenzaBase>) => void;
+    initialValues?: Partial<import("@/types/models/ricorrenza").RicorrenzaBase>;
+    onChangeForm?: (data: Partial<import("@/types/models/ricorrenza").RicorrenzaBase>) => void;
 };
 
 export type NewRicorrenzaButtonProps = {
     label?: string;
-    onSuccess?: (newRicorrenza: import("@/types/types/ricorrenza").Ricorrenza) => void;
+    onSuccess?: (newRicorrenza: import("@/types/models/ricorrenza").Ricorrenza) => void;
 };
