+++ Frontend-nextjs/types/panoramica/components/index.ts
@@ -1,4 +1,4 @@
-import type { Transaction } from "@/types/types/transaction";
+import type { Transaction } from "@/types/models/transaction";
 
 export type CalendarGridProps = {
     transactions: Transaction[];
